refactor: move audit package to interanl
