+++ sdk/openapi.go
@@ -15,8 +15,8 @@
 package sdk
 
 import (
-	"github.com/rond-authz/rond/audit"
 	"github.com/rond-authz/rond/core"
+	"github.com/rond-authz/rond/internal/audit"
 	"github.com/rond-authz/rond/metrics"
 	"github.com/rond-authz/rond/openapi"
